[refac] Refactor S3 package structure to use 'aws.s3' namespace
