+++ edukit-core/src/main/java/com/edukit/core/notice/facade/NoticeFacade.java
@@ -8,8 +8,8 @@ import com.edukit.core.notice.facade.response.NoticeGetResponse;
 import com.edukit.core.notice.facade.response.NoticeResponse;
 import com.edukit.core.notice.facade.response.NoticesGetResponse;
 import com.edukit.core.notice.service.NoticeService;
-import com.edukit.external.s3.S3Service;
-import com.edukit.external.s3.response.UploadPresignedUrlResponse;
+import com.edukit.external.aws.s3.S3Service;
+import com.edukit.external.aws.s3.response.UploadPresignedUrlResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
