+++ edukit-core/src/main/java/com/edukit/core/notice/facade/response/NoticeFileUploadPresignedUrlCreateResponse.java
@@ -1,7 +1,7 @@
 package com.edukit.core.notice.facade.response;
 
 import com.edukit.core.notice.entity.NoticeFile;
-import com.edukit.external.s3.response.UploadPresignedUrlResponse;
+import com.edukit.external.aws.s3.response.UploadPresignedUrlResponse;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
