+++ edukit-external/src/main/java/com/edukit/external/aws/s3/S3Service.java
@@ -1,8 +1,8 @@
-package com.edukit.external.s3;
+package com.edukit.external.aws.s3;
 
-import com.edukit.external.s3.exception.S3ErrorCode;
-import com.edukit.external.s3.exception.S3Exception;
-import com.edukit.external.s3.response.UploadPresignedUrlResponse;
+import com.edukit.external.aws.s3.exception.S3ErrorCode;
+import com.edukit.external.aws.s3.exception.S3Exception;
+import com.edukit.external.aws.s3.response.UploadPresignedUrlResponse;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
